// src/client/types.gen.ts
export type EmailRegisterIn = {
    /**
     * Email
     */
    email: string;
    /**
     * Password
     */
    password: string;
};

/**
 * EmailVerifyConfirmIn
 *
 * 验证邮箱请求体
 */
export type EmailVerifyConfirmIn = {
    /**
     * Email
     */
    email: string;
    /**
     * Otp
     *
     * 验证码
     */
    otp: string;
};

/**
 * EmailVerifyRequestIn
 *
 * 邮箱验证请求体
 */
export type EmailVerifyRequestIn = {
    /**
     * Email
     */
    email: string;
};
// ...
export type LoginIn = {
    /**
     * Email
     *
     * 登录邮箱
     */
    email: string;
    /**
     * Password
     *
     * 密码
     */
    password: string;
};
// ...
export type RegisterByEmailData = {
    body: EmailRegisterIn;
    path?: never;
    query?: never;
    url: '/api/v1/auth/register/email';
};
// ...
export type LoginData = {
    body: LoginIn;
    path?: never;
    query?: never;
    url: '/api/v1/auth/login';
};
// ...
export type RequestEmailVerificationData = {
    body: EmailVerifyRequestIn;
    path?: never;
    query?: never;
    url: '/api/v1/auth/email-verification/request';
};
// ...
export type VerifyEmailData = {
    body: EmailVerifyConfirmIn;
    path?: never;
    query?: never;
    url: '/api/v1/auth/email-verification/verify';
};